backend/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.1.0
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
greenlet==1.0.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyrsistent==0.17.3
python-dotenv==0.17.0
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
